finding the height

+++ 9-binary_tree_height.c
@@ -7,7 +7,7 @@
 
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int left_height, right_height;
+	size_t left_height, right_height;
 
 	if (tree == NULL)
 		return (-1);
